Separate code out into classes
Having all of the code in one big class makes it harder to figure out
what to change. Having separate classes should also make it easier to
write tests when I get round to that.
