+++ lib/everypolitician_data_pull_request.rb
@@ -1,0 +1,43 @@
+require 'github'
+
+class EverypoliticianDataPullRequest
+  include Github
+
+  attr_reader :pull_request
+
+  def initialize(pull_request_number)
+    @pull_request = github.pull(everypolitician_data_repo, pull_request_number)
+  end
+
+  def updated_countries_json?
+    files = github.pull_files(everypolitician_data_repo, pull_request[:number])
+    files.map { |f| f[:filename] }.flatten.uniq.include?('countries.json')
+  end
+
+  def list_of_commit_messages
+    commits = github.pull_commits(
+      everypolitician_data_repo,
+      pull_request[:number]
+    )
+    messages = commits.map do |commit|
+      commit.commit.message.lines.first.chomp
+    end
+    messages.map { |m| "- #{m}" }.join("\n")
+  end
+
+  def html_url
+    pull_request[:html_url]
+  end
+
+  def title
+    pull_request[:title]
+  end
+
+  def number
+    pull_request[:number]
+  end
+
+  def everypolitician_data_repo
+    @everypolitician_data_repo ||= ENV.fetch('EVERYPOLITICIAN_DATA_REPO', 'everypolitician/everypolitician-data')
+  end
+end
