+++ lib/github.rb
@@ -1,0 +1,5 @@
+module Github
+  def github
+    @github ||= Octokit::Client.new(access_token: ENV['GITHUB_ACCESS_TOKEN'])
+  end
+end
